refactor: move format_title to utils.py
