+++ faugus/shortcut.py
@@ -50,12 +50,6 @@ desktop_dir = PathManager.user_desktop()
 
 _ = setup_gettext('faugus-launcher')
 
-def format_title(title):
-    title = title.strip().lower()
-    title = re.sub(r"[^\w\s-]", "", title)
-    title = re.sub(r"\s+", "-", title)
-    return title
-
 class CreateShortcut(Gtk.Window, HiDpiMixin):
     def __init__(self, file_path):
         super().__init__(title="Faugus Launcher")
